% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_auth_log_out}
\alias{envrsk_auth_log_out}
\title{Function to log out from the API}
\usage{
envrsk_auth_log_out()
}
\value{
No return value. This function clears certain variables and environment variables.
}
\description{
This function logs out from the API by removing the environment variables used.
}
\examples{
\dontrun{
envrsk_auth_log_out()
}
}
