% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_auth_renew_access_token}
\alias{envrsk_auth_renew_access_token}
\title{Function to renew access token for API authentication}
\usage{
envrsk_auth_renew_access_token(force_renew = FALSE)
}
\arguments{
\item{force_renew}{A boolean value indicating whether to force the renewal
of the access token. Default is FALSE.}
}
\value{
A message indicating the validity of the access token or any error message.
}
\description{
This function renews the access token needed for API authentication.
It optionally forces a token renewal. If user ID and password are
available in the system environment variables, it uses them to get
a new access token. If the current access token exists and has not
expired, it continues to use it unless force_renew is TRUE.
}
\examples{
\dontrun{
envrsk_auth_renew_access_token(force_renew = TRUE)
}
}
