% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_auth_set_access_token}
\alias{envrsk_auth_set_access_token}
\title{Function to set access token for API authentication}
\usage{
envrsk_auth_set_access_token()
}
\value{
No return value. This function prints a message if the access token is
successfully retrieved.
}
\description{
This function sets the environment variables for user ID and password,
and retrieves an access token for API authentication. It uses the user's
input for the email and password as credentials for authentication.
The access token retrieved is stored globally in the 'my_access_token' variable.
}
\examples{
\dontrun{
envrsk_auth_set_access_token()
}
}
