% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_auth_set_access_token_non_interactively}
\alias{envrsk_auth_set_access_token_non_interactively}
\title{Set Access Token Non-Interactively}
\usage{
envrsk_auth_set_access_token_non_interactively(usr_id, usr_pwd)
}
\arguments{
\item{usr_id}{The user ID for authentication.}

\item{usr_pwd}{The password for authentication.}
}
\value{
A list containing the status code and a message. If the authentication is successful,
the list will have a status code of 200 and a message indicating the token has been acquired.
If the authentication fails, it returns a status code of 400 along with the error message.
}
\description{
This function sets the access token and its expiry as environment variables after
authenticating a user with the provided user ID and password. The function retrieves
an access token and checks the status of the
retrieval. If successful, the access token and expiry are set as environment variables.
If not successful, it returns an error message with the status code.
}
\examples{
\dontrun{
envrsk_auth_set_access_token_non_interactively("your_username", "your_password")
}

}
