% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_decorate_portfolio_with_product_type}
\alias{envrsk_decorate_portfolio_with_product_type}
\title{Decorate Portfolio with Product Type}
\usage{
envrsk_decorate_portfolio_with_product_type(positions, simplify = TRUE)
}
\arguments{
\item{positions}{A list of positions to be enriched with additional information.}

\item{simplify}{Logical, indicating whether to simplify the output. If TRUE, o
nly the output content is returned.
       If FALSE, the entire API response is returned. Default is TRUE.}
}
\value{
If the API call is successful, a list containing the enriched positions
is returned.
        If the API call is unsuccessful, the original API response is returned.
}
\description{
This function calls the 'decorate-table-id' endpoint of the EnvisionRisk API
to enrich portfolio position data with additional information based on position ID.
}
\examples{
\dontrun{
dt_positions_without_product_type <- as.data.frame(list("symbol"   = c("AAPL.US", "DANSKE.CO",
                                                                       "CashUSD", "AGG.US"),
                                   "quantity" = c(129, 768, 69000, 89)))
dt_positions <- envrsk_decorate_with_product_type(positions = dt_positions_without_product_type)
}
}
