% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_decorate_portfolio_with_uid}
\alias{envrsk_decorate_portfolio_with_uid}
\title{Decorates the portfolio with IDs using envrsk API}
\usage{
envrsk_decorate_portfolio_with_uid(access_token, positions, simplify = TRUE)
}
\arguments{
\item{access_token}{Character. The access token used for the envrsk API
authentication.}

\item{positions}{Data Frame or List. The portfolio positions that should
be decorated. This input should align with the envrsk API requirements for
the `decorate-table-id` endpoint.}

\item{simplify}{Logical. If TRUE, the function only returns the 'Output' field
of the API response as a data.table. If FALSE, the function returns the
entire response object. Default is TRUE.}
}
\value{
If the API call is successful (HTTP status code 200) and `simplify = TRUE`,
a data.table from the 'Output' field of the API response is returned. If
`simplify = FALSE`, the entire response object is returned. If the API call
is not successful, the function returns the API response with an error message.
}
\description{
This function uses the envrsk API to decorate a given portfolio with the
respective ID. The decoration process typically includes the addition of
meta-data, additional risk factors or calculations based on the portfolio
positions.
}
\examples{
\dontrun{
dt_positions <- as.data.frame(list("symbol"        = c("AAPL.US", "DANSKE.CO",
                                                       "CashUSD", "AGG.US"),
                                   "position_type" = c("single_stock",
                                                       "single_stock", "cash", "etf"),
                                   "quantity"      = c(129, 768, 69000, 89)))

dt_positions <- envrsk_decorate_portfolio_with_uid(dt_positions)
}
}
