% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_get}
\alias{envrsk_get}
\title{Function to make a GET request to an API}
\usage{
envrsk_get(access_token, url, params)
}
\arguments{
\item{access_token}{A string containing the access token.}

\item{url}{A string containing the URL of the API endpoint.}

\item{params}{A list containing the query parameters for the GET request.}
}
\value{
A list containing the status code and the content of the API response.
}
\description{
This function sends a GET request to the specified API URL.
It requires an access token, a URL, and a list of query parameters.
It adds the access token to the headers of the request and sends
the GET request using the 'httr' package.
}
\keyword{call}
\keyword{endpoint}
\keyword{headers}
\keyword{internal}
\keyword{post}
\keyword{the}
\keyword{to}
