% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_get_manifest}
\alias{envrsk_get_manifest}
\title{Get Manifest}
\usage{
envrsk_get_manifest()
}
\value{
If the API call is successful, a list containing the manifest data is
returned,
        with portfolio constituents combined into a single vector.
        If the API call is unsuccessful, the original API response is returned.
}
\description{
This function calls the 'get-manifest' endpoint of the EnvisionRisk API to
obtain the manifest, which includes available parameters for portfolio construction
and risk computation.
}
\examples{
\dontrun{
manifest <- envrsk_get_manifest()
}
}
