% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_instrument_delta_vector}
\alias{envrsk_instrument_delta_vector}
\title{Instrument Delta Vector}
\usage{
envrsk_instrument_delta_vector(
  date,
  symbols,
  base_cur = NULL,
  horizon = NULL,
  volatility_id = NULL
)
}
\arguments{
\item{date}{A string, the date at which the delta vector is to be computed (YYYY-MM-DD).}

\item{symbols}{A list, the symbols of the financial instruments for which the
delta vector is to be computed.}

\item{base_cur}{A string, the base currency in which the delta vector is to be computed.}

\item{horizon}{An integer, the time horizon over which the delta vector is computed.}

\item{volatility_id}{An integer, the ID of the volatility model to be used.}
}
\value{
A list containing the input parameters, the time of operation,
        the output delta vector data, and details of mapped and unmapped symbols.
        If the API call is unsuccessful, the original API response is returned.
}
\description{
This function calls the 'instrument-delta-vector' endpoint of the EnvisionRisk API
to compute the delta vector for a set of financial instruments. The delta vector
provides
the partial derivatives of the financial instruments' value with respect to an
underlying
asset price, providing an indication of the sensitivity of the instruments' value to changes
in the asset price.
}
\examples{
\dontrun{
delta_vec <- envrsk_instrument_delta_vector(date = "2023-06-01",
symbols = c("AAPL", "GOOGL"))
}
}
