% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_instrument_expected_shortfall}
\alias{envrsk_instrument_expected_shortfall}
\title{Instrument Expected Shortfall}
\usage{
envrsk_instrument_expected_shortfall(
  date,
  symbols,
  base_cur = NULL,
  horizon = NULL,
  signif_level = NULL,
  volatility_id = NULL
)
}
\arguments{
\item{date}{A string, the date at which ES is to be computed (YYYY-MM-DD).}

\item{symbols}{A list, the symbols of the financial instruments for which ES
is to be computed.}

\item{base_cur}{A string, the base currency in which ES is to be computed.}

\item{horizon}{An integer, the time horizon over which ES is computed.}

\item{signif_level}{A numeric, the significance level for the ES calculation
(between 0 and 1).}

\item{volatility_id}{An integer, the ID of the volatility model to be used.}
}
\value{
A list containing the input parameters, the time of operation,
        the output ES data, and details of mapped and unmapped symbols.
        If the API call is unsuccessful, the original API response is returned.
}
\description{
This function calls the 'instrument-expected-shortfall' endpoint of the EnvisionRisk
API
to compute the Expected Shortfall (ES), also known as Conditional Value at Risk
(CVaR), for a set of financial instruments.
ES is a risk measure that quantifies the expected value of loss given that an
event beyond the VaR threshold has occurred.
}
\examples{
\dontrun{
es <- envrsk_instrument_expected_shortfall(date = "2023-06-01",
                                           symbols = c("AAPL.US", "DANSKE.CO"))
}
}
