% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_instrument_performance_raw}
\alias{envrsk_instrument_performance_raw}
\title{Instrument Performance Raw}
\usage{
envrsk_instrument_performance_raw(
  symbols,
  from = NULL,
  to = NULL,
  days = 1,
  direction = "lead",
  overlap = TRUE
)
}
\arguments{
\item{symbols}{A list, the symbols of the financial instruments for which performance
data is required.}

\item{from}{A string, the start date for the performance period (YYYY-MM-DD).}

\item{to}{A string, the end date for the performance period (YYYY-MM-DD).}

\item{days}{An integer, the frequency at which performance data is calculated.}

\item{direction}{A string, whether the performance is calculated in a leading
('lead') or lagging ('lag') manner.}

\item{overlap}{A boolean, whether overlapping returns are allowed in the calculation.}
}
\value{
A list containing the input parameters, the time of operation,
        the output performance data, and details of mapped and unmapped symbols.
        If the API call is unsuccessful, the original API response is returned.
}
\description{
This function calls the 'instrument-performance-raw' endpoint of the EnvisionRisk API
to retrieve performance data for a set of financial instruments. The endpoint does not
apply currency conversion. The performance is calculated over a specified period
and frequency.
}
\examples{
\dontrun{
performance <- envrsk_instrument_performance_raw(symbols = c("AAPL.US", "DANSKE.CO"))
}
}
