% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_instrument_search}
\alias{envrsk_instrument_search}
\title{Instrument Search}
\usage{
envrsk_instrument_search(
  partial_name = NULL,
  partial_symbol = NULL,
  partial_exchange_id = NULL,
  position_type = NULL,
  valid_at = NULL
)
}
\arguments{
\item{partial_name}{A string, a partial name of the financial instrument.}

\item{partial_symbol}{A string, a partial symbol of the financial instrument.}

\item{partial_exchange_id}{A string, a partial exchange ID of the financial instrument.}

\item{position_type}{A string, the type of position to search for.}

\item{valid_at}{A string, date at which the financial instrument must be valid (YYYY-MM-DD).}
}
\value{
A data frame containing the details of the instruments matching the search criteria
        or the original API response if the status code is not 200.
}
\description{
This function calls the 'search-instrument' endpoint of the EnvisionRisk API
to search for financial instruments by partial name, symbol, exchange ID, or
position type.
The search is done based on the current or specified date.
}
\examples{
\dontrun{
instruments <- envrsk_instrument_search(partial_name = "AAPL")
}
}
