% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_instrument_value_at_risk_raw}
\alias{envrsk_instrument_value_at_risk_raw}
\title{Instrument Value at Risk}
\usage{
envrsk_instrument_value_at_risk_raw(
  date,
  symbols,
  horizon = NULL,
  signif_level = NULL,
  volatility_id = NULL
)
}
\arguments{
\item{date}{A string, the date at which VaR is to be computed (YYYY-MM-DD).}

\item{symbols}{A list, the symbols of the financial instruments for which VaR
is to be computed.}

\item{horizon}{An integer, the time horizon over which VaR is computed.}

\item{signif_level}{A numeric, the significance level for the VaR calculation
(between 0 and 1).}

\item{volatility_id}{An integer, the ID of the volatility model to be used.}
}
\value{
A list containing the input parameters, the time of operation,
        the output VaR data, and details of mapped and unmapped symbols.
        If the API call is unsuccessful, the original API response is returned.
}
\description{
This function calls the 'instrument-value-at-risk-raw' endpoint of the EnvisionRisk API
to compute the Value at Risk (VaR) for a set of financial instruments.
VaR is a statistical measure that quantifies the level of financial risk within a
firm or investment portfolio over a specific time frame.
}
\examples{
\dontrun{
var <- envrsk_instrument_value_at_risk_raw(access_token = "your_access_token",
                                           date = "2023-06-01",
                                           symbols = c("AAPL.US", "DANSKE.CO"))
}
}
