% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_manifest_restore_to_default}
\alias{envrsk_manifest_restore_to_default}
\title{envrsk_manifest_restore_to_default}
\usage{
envrsk_manifest_restore_to_default()
}
\value{
If the status code of the response is 200, the function will not
return a value but print a message: "OK - Manifest updated".
If the status code is anything other than 200, the function will
return the complete response object.
}
\description{
This function restores the default configuration of the manifest
by sending a POST request to a specific API endpoint.
If the status code of the response is 200, it prints a success message.
Otherwise, it returns the whole response.
}
\details{
This function constructs the API url using the `get_api_url` helper function
with "restore-manifest" as the endpoint.
It then prepares the query parameters and removes any parameters
with zero length by indexing with `lengths(.params) != 0`.
It sends a POST request to the `api_url` using the `envrsk_post` helper function.
The function requires no explicit arguments from the user, but utilizes
an access token generated by the `get_access_token` helper function.
It assumes a default body content (`list(manifest)`).
}
\examples{
\dontrun{
  envrsk_manifest_restore_to_default()
}

}
