% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_market_stress_volatility}
\alias{envrsk_market_stress_volatility}
\title{Market Stress Volatility}
\usage{
envrsk_market_stress_volatility(symbols)
}
\arguments{
\item{symbols}{A list, the symbols of the financial instruments for which the
market stress volatilities are to be obtained.}
}
\value{
A list containing the input parameters, the time of operation, the output
market stress volatility data,
        and details of mapped and unmapped symbols. If the API call is unsuccessful,
        the original API response is returned.
}
\description{
This function calls the 'market-stress-volatility' endpoint of the EnvisionRisk
API to
obtain the market stress volatilities for a set of financial instruments.
}
\examples{
\dontrun{
stress_volatilities <- envrsk_market_stress_volatility(symbols = c("AAPL.US", "DANSKE.CO"))
}
}
