% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_market_volatility}
\alias{envrsk_market_volatility}
\title{Market Volatility}
\usage{
envrsk_market_volatility(symbols)
}
\arguments{
\item{symbols}{A list, the symbols of the financial instruments for which the
market volatilities are to be obtained.}
}
\value{
A list containing the input parameters, the time of operation, the output
market volatility data,
        and details of mapped and unmapped symbols. If the API call is unsuccessful,
        the original API response is returned.
}
\description{
This function calls the 'market-volatility' endpoint of the EnvisionRisk API to
obtain the market volatilities for a set of financial instruments.
}
\examples{
\dontrun{
market_volatilities <- envrsk_market_volatility(symbols = c("AAPL.US", "DANSKE.CO"))
}
}
