% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_portfolio_delta_vector}
\alias{envrsk_portfolio_delta_vector}
\title{Function to calculate the delta vector of a portfolio}
\usage{
envrsk_portfolio_delta_vector(
  date,
  positions,
  base_cur = NULL,
  horizon = NULL,
  signif_level = NULL,
  volatility_id = NULL,
  report_depth = NULL,
  simplify = FALSE
)
}
\arguments{
\item{date}{A string containing the date for the calculation.}

\item{positions}{A list containing the positions in the portfolio.}

\item{base_cur}{An optional string containing the base currency.}

\item{horizon}{An optional numeric containing the horizon.}

\item{signif_level}{An optional numeric containing the significant level.}

\item{volatility_id}{An optional string containing the volatility ID.}

\item{report_depth}{An optional numeric containing the report depth.}

\item{simplify}{An optional logical value indicating whether to simplify the
return values.}
}
\value{
A list containing the API response.
}
\description{
This function sends a POST request to the 'portfolio-delta-vector' API endpoint
to calculate the delta vector of a portfolio on a specified date.
It requires an access token, a date, and the positions in the portfolio.
It also allows for optional parameters like base currency, horizon,
volatility ID, and report depth. It returns the response from the API.
}
\examples{
\dontrun{
dt_positions <- as.data.frame(list("symbol"        = c("AAPL.US", "DANSKE.CO",
"CashUSD", "AGG.US"),
                                   "position_type" = c("single_stock",
                                   "single_stock", "cash", "etf"),
                                   "quantity"      = c(129, 768, 69000, 89)))

delta_vector_1 <- envrsk_portfolio_delta_vector(date         = "2022-12-31",
                                                positions    = dt_positions)

delta_vector_2 <- envrsk_portfolio_delta_vector(date          = "2022-12-31",
                                                positions     = dt_positions,
                                                base_cur      = "USD",
                                                horizon       = 1,
                                                volatility_id = "point_in_time",
                                                report_depth  = 0,
                                                simplify      = TRUE)
}
}
