% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_portfolio_economic_capital_regular}
\alias{envrsk_portfolio_economic_capital_regular}
\title{Function to calculate the economic capital of a portfolio}
\usage{
envrsk_portfolio_economic_capital_regular(
  date,
  positions,
  base_cur = NULL,
  horizon = NULL,
  signif_level = NULL,
  volatility_id = NULL,
  expected_roe = NULL,
  report_depth = NULL,
  simplify = FALSE
)
}
\arguments{
\item{date}{A string containing the date for the calculation.}

\item{positions}{A list containing the positions in the portfolio.}

\item{base_cur}{An optional string containing the base currency.}

\item{horizon}{An optional numeric containing the horizon.}

\item{signif_level}{An optional numeric containing the significant level.}

\item{volatility_id}{An optional string containing the volatility ID.}

\item{expected_roe}{An optional numeric containing the expected Return on
Equity (ROE).}

\item{report_depth}{An optional numeric containing the report depth.}

\item{simplify}{An optional logical value indicating whether to simplify the
return values.

Economic capital in market risk management is the amount of risk capital a
company holds to withstand unexpected losses due to market risk. It's essentially
a safety net that protects the company's financial health from severe market
downturns. It's computed based on statistical measure Expected Shortfall (ES),
which estimate potential losses under extreme but plausible scenarios. A higher
economic capital indicates a larger buffer against market volatility but also
may suggest greater exposure to risk. It's crucial in risk management, capital
allocation, and strategic decision-making.}
}
\value{
A list containing the API response.
}
\description{
This function sends a POST request to the 'portfolio-economic-capital-regular'
API endpoint
to calculate the economic capital of a portfolio on a specified date.
It requires an access token, a date, and the positions in the portfolio.
It also allows for optional parameters like base currency, horizon,
volatility ID, expected ROE (Return on Equity), and report depth.
It returns the response from the API.
}
\examples{
\dontrun{
dt_positions <- as.data.frame(list("symbol"        = c("AAPL.US", "DANSKE.CO",
"CashUSD", "AGG.US"),
                                   "position_type" = c("single_stock",
                                   "single_stock", "cash", "etf"),
                                   "quantity"      = c(129, 768, 69000, 89)))

economic_capital_1 <- envrsk_portfolio_economic_capital_regular(date         = "2022-12-31",
                                                                positions    = dt_positions)

economic_capital_2 <- envrsk_portfolio_economic_capital_regular(date          = "2022-12-31",
                                                                positions     = dt_positions,
                                                                base_cur      = "USD",
                                                                horizon       = 10,
                                                                signif_level  = 0.99,
                                                                volatility_id = "downturn",
                                                                expected_roe  = 0.10,
                                                                report_depth  = 0,
                                                                simplify      = TRUE)
}
}
