% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_portfolio_portfolio_hyp_rskadj_perf_component}
\alias{envrsk_portfolio_portfolio_hyp_rskadj_perf_component}
\title{Compute Portfolio Hypothetical Risk Adjusted Performance (Component)}
\usage{
envrsk_portfolio_portfolio_hyp_rskadj_perf_component(
  date,
  positions,
  base_cur = NULL,
  horizon = NULL,
  signif_level = NULL,
  volatility_id = NULL,
  expected_roe = NULL,
  report_depth = NULL,
  simplify = FALSE
)
}
\arguments{
\item{date}{A string in the format 'yyyy-mm-dd', required for the API
call. The date of the portfolio.}

\item{positions}{A data frame or matrix. Each row represents a position
in the portfolio with necessary details.}

\item{base_cur}{A string representing the base currency for the portfolio.
Default is NULL.}

\item{horizon}{Numeric. The time horizon in days for which the value at
risk is calculated. Default is NULL.}

\item{signif_level}{Numeric. The significance level for the value at risk
calculation. Default is NULL.}

\item{volatility_id}{An optional identifier for a specific volatility model
to use in the calculation. Default is NULL.}

\item{expected_roe}{Expected return on equity, as a decimal. Default is NULL.}

\item{report_depth}{The depth to which the report should calculate risk. Default
is NULL.}

\item{simplify}{Logical indicating whether the result should be simplified,
if possible. Default is FALSE.}
}
\value{
If simplify = FALSE, a list with the following components is returned:
        - "Input": A list containing the input parameters used in the API call.
        - "tech_opr": The timestamp of when the API call was made.
        - "Output": A data.table with the calculated risk performance.
        - "symbols_mapped": A data.table with the symbol mapping.
        - "symbols_unmapped": A data.table with the symbols that could not be mapped.
        If simplify = TRUE, only the "Output" data.table is returned.
}
\description{
This function communicates with a given API endpoint to compute and return the
hypothetical risk adjusted performance of a given portfolio,
considering positions provided.
}
\examples{
\dontrun{
dt_positions <- as.data.frame(list("symbol"        = c("AAPL.US", "DANSKE.CO",
"CashUSD", "AGG.US"),
                                   "position_type" = c("single_stock",
                                   "single_stock", "cash", "etf"),
                                   "quantity"      = c(129, 768, 69000, 89)))

result <- envrsk_portfolio_portfolio_hyp_rskadj_perf_component(date         = "2023-05-20",
                                                               positions    = dt_positions,
                                                               base_cur     = "USD",
                                                               horizon      = 1,
                                                               signif_level = 0.95,
                                                               expected_roe = 0.1,
                                                               report_depth = 3,
                                                               simplify = TRUE)
}
}
