% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_portfolio_risk_component}
\alias{envrsk_portfolio_risk_component}
\title{Function for estimating risk components of a portfolio}
\usage{
envrsk_portfolio_risk_component(
  date,
  positions,
  base_cur = NULL,
  horizon = NULL,
  signif_level = NULL,
  volatility_id = NULL,
  report_depth = NULL,
  simplify = FALSE
)
}
\arguments{
\item{date}{A date for which the portfolio risk should be estimated.}

\item{positions}{A list of positions in the portfolio.}

\item{base_cur}{An optional base currency for the portfolio. Default is NULL.}

\item{horizon}{An optional time horizon for the risk estimate. Default is NULL.}

\item{signif_level}{An optional significance level for the risk measure.
Default is NULL.}

\item{volatility_id}{An optional volatility ID. Default is NULL.}

\item{report_depth}{An optional depth of the report. Default is NULL.}

\item{simplify}{A flag indicating whether to simplify the output. Default is FALSE.

Value at Risk (VaR) is a statistical measure that estimates the maximum potential
loss over a specified time horizon at a given confidence level. Expected
Shortfall (ES), also known as Conditional Value at Risk (CVaR), estimates the
average loss in the event that the VaR is exceeded. Both are forward-looking
as they use statistical analysis based on historical data and volatility to
predict potential future losses. They are essential for market risk management,
helping organizations understand their risk exposure and potential financial
impact under adverse market conditions.

Component Value at Risk (ComponentVaR) and Component Expected Shortfall
(ComponentES) are measures that break down the total VaR and ES of a portfolio
into individual components corresponding to each asset in the portfolio.

ComponentVaR for a particular asset represents the contribution of that asset
to the portfolio's overall VaR. Similarly, CES quantifies an individual asset's
contribution to the portfolio's total expected shortfall.

These measures are crucial for market risk management as they provide detailed
insights into the risk contribution of each asset in the portfolio. By
identifying the assets that contribute most to the portfolio's risk, a risk
manager can better optimize the portfolio's risk-return tradeoff. For example,
they may decide to reduce the portfolio's exposure to assets with high CVaR or
CES, diversify the asset allocation, or use hedging strategies to mitigate risk.}
}
\value{
A processed portfolio return value.
}
\description{
This function uses the 'portfolio-risk-regular' API endpoint to estimate
portfolio risk. It requires an access token and a date, along with a
list of positions. It optionally takes in a base currency, a horizon,
a significance level, a volatility ID, a report depth, and a flag
to simplify the output.
}
\examples{
\dontrun{
dt_positions <- as.data.frame(list("symbol"        = c("AAPL.US", "DANSKE.CO",
"CashUSD", "AGG.US"),
                                   "position_type" = c("single_stock",
                                   "single_stock", "cash", "etf"),
                                   "quantity"      = c(129, 768, 69000, 89)))
result_1 <- envrsk_portfolio_risk_component(date         = "2022-12-31",
                                          positions    = dt_positions)

result_2 <- envrsk_portfolio_risk_component(date          = "2022-12-31",
                                          positions     = dt_positions,
                                          base_cur      = "USD",
                                          horizon       = 1,
                                          signif_level  = 0.975,
                                          volatility_id = "point_in_time",
                                          report_depth  = 0,
                                          simplify      = TRUE)
}
}
