% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_post}
\alias{envrsk_post}
\title{Function to make a POST request to an API}
\usage{
envrsk_post(access_token, url, body, params)
}
\arguments{
\item{access_token}{A string containing the access token.}

\item{url}{A string containing the URL of the API endpoint.}

\item{body}{A list containing the body of the POST request.}

\item{params}{A list containing the query parameters for the POST request.}
}
\value{
A list containing the status code and the content of the API response.
}
\description{
This function sends a POST request to the specified API URL.
It requires an access token, a URL, a body for the POST request,
and a list of query parameters. It adds the access token to the
headers of the request and sends the POST request using the 'httr' package.
}
\keyword{call}
\keyword{endpoint}
\keyword{headers}
\keyword{internal}
\keyword{post}
\keyword{the}
\keyword{to}
