% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_update_manifest}
\alias{envrsk_update_manifest}
\title{Update a Manifest}
\usage{
envrsk_update_manifest(manifest)
}
\arguments{
\item{manifest}{A list, data frame or other structure that holds the manifest
to be updated. This must match the format expected by the API.}
}
\value{
NULL If the API request is successful (i.e., HTTP status code 200),
   this function prints a success message and returns NULL. If the API request
   is not successful, this function returns the full response from the API.
}
\description{
This function makes an HTTP POST request to the 'put-manifest' API endpoint
to update the specified manifest.
}
\examples{
\dontrun{
  # Update a manifest with data
  manifest <- list(id = "abc123", status = "complete")
  envrsk_update_manifest(manifest)
}

\dontrun{
my_manifest              <- envrsk_get_manifest()
my_manifest$BASE_CUR     <- "DKK"
my_manifest$SIGNIF_LEVEL <- 0.975

envrsk_update_manifest(my_manifest)
}
}
