% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_workflow_risk_snapshot}
\alias{envrsk_workflow_risk_snapshot}
\title{Run Workflow Risk Snapshot}
\usage{
envrsk_workflow_risk_snapshot(
  date,
  positions,
  base_cur = NULL,
  horizon = NULL,
  signif_level = NULL,
  volatility_id = NULL,
  report_depth = NULL,
  simplify = FALSE
)
}
\arguments{
\item{date}{The date for which to run the risk snapshot.}

\item{positions}{The positions data.}

\item{base_cur}{The base currency to use for the risk calculations. Default is NULL.}

\item{horizon}{The horizon for the risk calculations. Default is NULL.}

\item{signif_level}{The significance level for the risk calculations. Default is NULL.}

\item{volatility_id}{The volatility id to use for the risk calculations. Default is NULL.}

\item{report_depth}{The depth of the report to be generated. Default is NULL.}

\item{simplify}{Logical. If TRUE, the output is simplified. Default is FALSE.}
}
\value{
A list containing the inputs, technical operations, positions, portfolio delta vector,
        portfolio risk, mapped positions, and unmapped positions.
}
\description{
This function runs a workflow risk snapshot which calculates risk measures for a portfolio of positions.
The function makes an API request to the specified endpoint and processes the returned data. It achieves
this by making a post request to the specified endpoint with a variety of parameters that allow the user
to customize the level of detail and perspective of the risk report. The function returns a list of various
outputs, including the original input data, a technical operations report, portfolio positions, a
portfolio delta vector, a portfolio risk report, and information on the mapping of positions.
}
\examples{
\dontrun{
dt_positions <- as.data.frame(list("symbol"        = c("AAPL.US", "DANSKE.CO",
                                                       "CashUSD", "AGG.US"),
                                   "position_type" = c("single_stock",
                                   "single_stock", "cash", "etf"),
                                   "quantity"      = c(129, 768, 69000, 89)))

response_risk_snapshot_1 <- envrsk_workflow_risk_snapshot(date      = Sys.Date(),
                                                        positions = dt_positions)

response_risk_snapshot_2 <- envrsk_workflow_risk_snapshot(date      = Sys.Date(),
                                                   positions    = dt_positions,
                                                   base_cur     = "USD",
                                                   horizon      = 10,
                                                   signif_level = 0.99,
                                                   volatility_id = "downturn",
                                                   simplify = TRUE)
}
}
