% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{envrsk_workflow_weight_2_quantities}
\alias{envrsk_workflow_weight_2_quantities}
\title{envrsk_workflow_weight_2_quantities function}
\usage{
envrsk_workflow_weight_2_quantities(
  dt_snapshot_weight,
  init_port_market_value,
  base_cur,
  is_wide = FALSE,
  to_date = NULL
)
}
\arguments{
\item{dt_snapshot_weight}{A data table that contains portfolio weights at a certain date.}

\item{init_port_market_value}{A numeric value that represents the initial market value of the portfolio.}

\item{base_cur}{A character string that specifies the base currency for the calculation.}

\item{is_wide}{Optional. A logical value that indicates whether the input data table is in 'wide' format. Default is FALSE.}

\item{to_date}{Optional. A character string or Date object representing the ending date for the calculation. Default is the current date.}
}
\value{
If the API call is successful (i.e., status code 200), a list containing 'Title', 'Input' (which includes
'PortfolioWeights', 'base_cur', 'signif_level'), 'TechOpr', 'Output' (which includes 'PortfolioEvents' and 'PortfolioQuantites'),
and 'UnMappedSymbols'. If the API call fails, it returns the original output from the API call which includes the status code and error message.
}
\description{
This function calculates quantities based on weights in a portfolio, it communicates with an external API and
processes the received data. It enables you to convert portfolio weights to quantities for each instrument.
}
\examples{
\dontrun{
# Download a test dataset with positions weights in a wide format.
dt_snapshot_weight <- readRDS(url("https://www.dropbox.com/scl/fi/6864xzl4fdw7hw7wu5m8t/portfolio_weights_wide.rds?rlkey=ra44iz6ruwmte4zyi8itkpoww&raw=true","rb"))

# Process the weights into daily position-quantities
result_weights_2_quantities <- envrsk_workflow_weight_2_quantities(
  dt_snapshot_weight     = dt_snapshot_weight,
  init_port_market_value = 1000000,
  base_cur               = "DKK",
  is_wide                = TRUE)
}
}
