% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envrsk_api_bridge_2_R.R
\name{process_portfolio_return_values}
\alias{process_portfolio_return_values}
\title{Process API Response}
\usage{
process_portfolio_return_values(res_out, simplify)
}
\arguments{
\item{res_out}{A list containing the API response.}

\item{simplify}{A logical value indicating whether to simplify the return values.}
}
\value{
A list or a data frame depending on the 'simplify' parameter.
        If the API response status code is not 200, it returns the original response.
}
\description{
This function processes the response from the portfolio risk API.
It binds rows of the 'Output', 'Positions_Mapped', and 'Positions_UnMapped' lists
into data frames using the 'rbindlist' function from the 'data.table' package.
If the 'simplify' parameter is TRUE, it returns only the 'Output' data frame;
otherwise, it returns the entire response.
}
\examples{
\dontrun{
processed_output <- process_portfolio_return_values(res_out = api_response, simplify = TRUE)
}
}
